Move exception message extraction out of rules
